+++ lib/include/args_handler.h
@@ -1,7 +1,9 @@
 #pragma once
 
+#include <cstddef>
+
 class ArgsHandler {
 public:
-    static void Handle(int argc, char* argv[]);
+    static void Handle(size_t argc, char* argv[]);
     static void WriteHelp();
 };
